[TJ&KEN]: moved update and delete logic to companyService

+++ src/main/java/com/tw/apistackbase/service/CompanyService.java
@@ -4,7 +4,13 @@ import com.tw.apistackbase.core.Company;
 import com.tw.apistackbase.repositories.CompanyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+
+import java.util.Optional;
 
 @Service
 public class CompanyService {
@@ -28,7 +34,27 @@ public class CompanyService {
         return companyRepository.save(company);
     }
 
-    public void deleteById(Long id) {
-        companyRepository.deleteById(id);
+    public ResponseEntity<String> deleteById(Long id) {
+        Optional<Company> companyToDelete = Optional.ofNullable(companyRepository.findOneById(id));
+
+        if(companyToDelete.isPresent()) {
+            companyRepository.deleteById(id);
+            return new ResponseEntity<>("Deleted ID "+id, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+    public ResponseEntity<Company> updateNameById(Long id, Company company) {
+        Optional<Company> companyToUpdate = Optional.ofNullable(companyRepository.findOneById(id));
+
+        if(companyToUpdate.isPresent()) {
+            companyToUpdate.get().setName(company.getName());
+            companyToUpdate.get().setProfile(company.getProfile());
+            companyToUpdate.get().setEmployees(company.getEmployees());
+
+            Company savedCompany = companyRepository.save(companyToUpdate.get());
+            return new ResponseEntity<>(savedCompany, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
